# app/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jazz-life",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "List Artists",
        "operationId": "list_artists_api_artists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ArtistRead_"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Upsert Artist",
        "description": "Upsert keyed on spotify_id; returns the existing row if present (200).\n\nPhase B-3 PR-2: called by RecordFormModal when a Spotify album is selected\nand the album's artist is not yet in the DB. Idempotent by design: a\nduplicate POST never returns 409.",
        "operationId": "upsert_artist_api_artists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/{spotify_id}": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Get Artist",
        "description": "Fetch a single artist, filling in image_url from Spotify when it is NULL.\n\nThe \"lazy photo hydration\" endpoint called when ArtistDetailModal opens.\nSpotify-side errors are swallowed in the service layer, so when no image\nis available the artist is returned with image_url=null (the UI falls back\nto initials).",
        "operationId": "get_artist_api_artists__spotify_id__get",
        "parameters": [
          {
            "name": "spotify_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spotify Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-follows": {
      "post": {
        "tags": [
          "user-follows"
        ],
        "summary": "Follow Artist",
        "description": "Add `artist_id` to the current user's follows.\n\nIdempotent because it uses `UserFollowRepository.upsert`:\n- already an active follow: overwrite is a no-op, 201 + existing artist\n- an archived row exists: reset archived_flag=false and re-follow\n- no row: new INSERT\n\n404 if the artist is not in the `artists` table. The UI is expected to\nupsert it first via `POST /api/artists` (passing the Spotify search result\nmetadata as-is).",
        "operationId": "follow_artist_api_user_follows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFollowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-follows/{artist_id}": {
      "delete": {
        "tags": [
          "user-follows"
        ],
        "summary": "Unfollow Artist",
        "description": "Unfollow the artist with this Spotify ID (soft delete: archived_flag=true).\n\nOnly sets archived_flag, so the \"previously followed\" history is kept.\nlist_artist_ids returns archived_flag=false rows only, so the next release\nsync excludes this artist.\n\n404 if there is nothing to remove; 204 if already archived (idempotent).",
        "operationId": "unfollow_artist_api_user_follows__artist_id__delete",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtistCreate": {
        "properties": {
          "spotify_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Spotify Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "spotify_dynamic",
              "manual"
            ],
            "title": "Source",
            "default": "spotify_dynamic"
          }
        },
        "type": "object",
        "required": [
          "spotify_id",
          "name"
        ],
        "title": "ArtistCreate",
        "description": "Create-artist request; upserts (returns the existing row if spotify_id exists).\n\nCalled by the Phase B-3 PR-2 RecordFormModal when a Spotify album is\nselected and the album's artist is not yet in the DB."
      },
      "ArtistRead": {
        "properties": {
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "spotify_dynamic",
              "manual"
            ],
            "title": "Source"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          }
        },
        "type": "object",
        "required": [
          "spotify_id",
          "name",
          "image_url",
          "source",
          "added_at"
        ],
        "title": "ArtistRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListResponse_ArtistRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtistRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListResponse[ArtistRead]"
      },
...
      "UserFollowCreate": {
        "properties": {
          "artist_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Artist Id"
          }
        },
        "type": "object",
        "required": [
          "artist_id"
        ],
        "title": "UserFollowCreate",
        "description": "Request body of `POST /api/user-follows`.\n\nThe artist must already exist in the `artists` table (FK constraint).\nThe UI is expected to reach a follow in three steps: Spotify artist\nsearch → `POST /api/artists` (upsert) → this endpoint."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}